Further progress on graceful shutdowns, but not yet complete. Sig quit still seems to interfere with socket communications

+++ src/hippogriff.coffee
@@ -25,6 +25,30 @@ exports.master = (path_to_worker, path_to_config) ->
   # Use user provided config for overrides
   config = load_config_file config, path_to_config
   
+  # Start unix server for workers to report in
+  net.createServer((socket) ->
+    socket.on "data", (data) ->
+      data = data.toString()
+      console.log data
+      lines = data.split '\n'
+      for line in lines
+        if line[0..16] == "HELLO MY NAME IS "
+          worker_socket = new net.Socket()
+          console.log workers[line[17..-1]].socket_path
+          worker_socket.connect(workers[line[17..-1]].socket_path, () ->
+            worker_socket.on 'data', (data) ->
+              data = data.toString()
+              console.log data
+              lines = data.split '\n'
+              for line in lines
+                if line[0..4] == "PONG "
+                  workers[line[5..-1]].status = "okay"
+                else if line[0..12] == "GOODBYE FROM "
+                  delete workers[line[13..-1]]
+            socket.on 'error', (e) -> console.err util.inspect e
+          )
+          workers[line[17..-1]].socket = worker_socket
+  ).listen(config.socket)
   
   # Start our engines....
   for i in [0...config.workers]
@@ -35,15 +59,15 @@ exports.master = (path_to_worker, path_to_config) ->
     setInterval (() ->
       new_workers = []
       for worker of workers
-        console.log workers[worker].socket.destroyed
-        if workers[worker].status == "concerned"
-          console.log "Worker #{worker} failed to respond, SIGKILLing..."
-          workers[worker].handle.kill("SIGKILL")
-          new_workers.push workers[worker]
-          delete workers[worker]
-        else if workers[worker].status == "okay"
-          workers[worker].socket.write 'PING\n' if workers[worker].socket?
-          workers[worker].status = "concerned"
+        if workers[worker].socket?
+          if workers[worker].status == "concerned"
+            console.log "Worker #{worker} failed to respond, SIGKILLing..."
+            workers[worker].handle.kill("SIGKILL")
+            new_workers.push workers[worker]
+            delete workers[worker]
+          else if workers[worker].status == "okay"
+            workers[worker].socket.write 'PING\n' if workers[worker].socket?
+            workers[worker].status = "concerned"
       for new_worker in new_workers
         startWorker(new_worker.config, new_worker.path_to_worker)
     ), config.checkInterval
@@ -72,14 +96,11 @@ exports.master = (path_to_worker, path_to_config) ->
     clearInterval(child_watcher_id)
     for worker of workers
       if workers[worker].socket?
+        console.log "Quitting #{worker}"
         workers[worker].status = "quitting"
-        # console.log workers[worker].socket.destroyed
-        console.log workers[worker].socket.write("GO AWAY #{worker}" + '\n')
-        # console.log util.inspect fs.statSync workers[worker].socket.server.path
-        # fs.statSync
-        buff = new Buffer "GO AWAY #{worker}" + '\n'
-        # fs.writeSync workers[worker].socket.fd, buff, 0, buff.length, 1
-        # process.nextTick(workers[worker].socket.write("GO AWAY #{worker}" + '\n'))
+        console.log util.inspect workers[worker].socket
+        workers[worker].socket.write("GO AWAY #{worker}" + '\n')
+        workers[worker].socket.end()
       else
         # No way to communicate
         workers[worker].handle.kill("SIGKILL")
@@ -125,25 +146,9 @@ startWorker = (config, path_to_worker) ->
     
     socket_counter++
     process.env['_HIPPOGRIFF_SOCKET'] = "/tmp/hippogriff.#{socket_counter}.sock"
+    process.env['_HIPPOGRIFF_MASTER_SOCKET'] = config.socket
     console.log "starting #{process.env['_HIPPOGRIFF_SOCKET']}"
     fs.unlinkSync process.env['_HIPPOGRIFF_SOCKET'] if path.existsSync process.env['_HIPPOGRIFF_SOCKET']
-    # socket = new net.Socket()
-    
-    net.createServer((socket) ->
-      socket.on 'data', (data) ->
-        data = data.toString()
-        console.log data
-        lines = data.split '\n'
-        for line in lines
-          if line[0..4] == "PONG "
-            workers[line[5..-1]].status = "okay"
-          else if line[0..12] == "GOODBYE FROM "
-            delete workers[line[13..-1]]
-          else if line[0..16] == "HELLO MY NAME IS "
-            workers[line[17..-1]].socket = socket
-  
-      socket.on 'error', (e) -> console.log util.inspect e
-    ).listen process.env['_HIPPOGRIFF_SOCKET']
     
     worker = spawn "#{config.exec_binary}", ["#{path_to_worker}"], {cwd:path.dirname path_to_worker}
     worker.stdout.on 'data', (data) -> process.stdout.write data
@@ -155,11 +160,13 @@ startWorker = (config, path_to_worker) ->
         if Object.keys(workers).length == 0
           process.exit 0
     console.log "Started #{worker.pid}, as #{config.exec_binary} #{path_to_worker}"
+    
     workers[worker.pid] = {
       status: "okay"
       handle: worker
       path_to_worker: path_to_worker
       config: config
+      socket_path: process.env['_HIPPOGRIFF_SOCKET']
       }
 
 load_config_file = (defaults, path_to_config) ->
